Organize tests and add test script
