+++ package.json
@@ -8,7 +8,8 @@
     "dev": "http-server .",
     "build": "node build.js",
     "server": "node multiplayer-server.js",
-    "simulate": "node simulate.js"
+    "simulate": "node simulate.js",
+    "test": "jest"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.24.0",
